requirements.txt
channels==4.0.0
daphne==4.0.0
django==5.0
django-restframework==3.14.0
django-cloudinary-storage==0.3.0
djangorestframework-simplejwt==5.2.2
ipython==8.19.0
pytest-django==4.7.0
pytest-factory==2.10.0
websockets==12.0
